fix: CodeQL format-string alerts + two real, separate dev-tool build breaks found along the way
CodeQL alerts #23/#24/#25/#26/#27: explicit casts on printf arguments in
xenia_probe/regbreak/ui_harness so each vararg's type exactly matches its
format specifier, regardless of typedef resolution.

While verifying those against a real compile, found two genuine,
pre-existing, unrelated build breaks (both fixed and build-verified,
full compile+link):

- tools/regbreak/main.cpp referenced 32-bit CONTEXT register fields
  (Eip/Esp/Eax/etc.) against the real x64 _CONTEXT struct -- did not
  compile at all against the current SDK. Rewritten to the real x64
  register set (Rip/Rsp/Rax/etc.), DumpProbe widened to DWORD64/%016llX.
  This tool is x86-era (built for the now-permanently-removed aim-assist
  investigation); its candidate offset table stays x86-era data, not
  reusable for x64 without fresh RE -- documented in place.
- tools/ui_harness's host_stubs.cpp hadn't been updated for symbols
  controller_input.cpp gained since (Controller_DetectGlyphStyle,
  FrameBenchmark_AddRumbleMs/AddPollThreadMs, g_modConfig) -- 4 unresolved
  externals at link time. Added real, documented stubs/definition.

+++ tools/ui_harness/main.cpp
@@ -230,7 +230,7 @@ DWORD WINAPI SourceWatcherThreadProc(LPVOID)
                 CloseHandle(pi.hProcess);
                 CloseHandle(pi.hThread);
                 if (exitCode == 0) printf("[ui_harness] rebuild OK -- will hot-swap within ~500ms\n");
-                else printf("[ui_harness] rebuild FAILED (exit %lu) -- keeping last good version\n", exitCode);
+                else printf("[ui_harness] rebuild FAILED (exit %lu) -- keeping last good version\n", static_cast<unsigned long>(exitCode));
             } else {
                 printf("[ui_harness] failed to launch MSBuild -- check the hardcoded path in SourceWatcherThreadProc\n");
             }
